完成用户注册、登录、鉴权等功能

+++ service/user_service/query_user_info.go
@@ -6,17 +6,17 @@ import (
 )
 
 type UserInfo struct {
-	ID              uint   `json:"id"`
+	ID              int64  `json:"id"`
 	Name            string `json:"name"`
 	Avatar          string `json:"avatar"`
 	BackgroundImage string `json:"background_image"`
 	Signature       string `json:"signature"`
-	FollowingCount  uint   `json:"follow_count"`
-	FollowerCount   uint   `json:"follower_count"`
+	FollowingCount  int64  `json:"follow_count"`
+	FollowerCount   int64  `json:"follower_count"`
 	IsFollow        bool   `json:"is_follow"`
-	TotalFavorited  uint   `json:"total_favorited"`
-	WorkCount       uint   `json:"work_count"`
-	FavoriteCount   uint   `json:"favorite_count"`
+	TotalFavorited  int64  `json:"total_favorited"`
+	WorkCount       int64  `json:"work_count"`
+	FavoriteCount   int64  `json:"favorite_count"`
 }
 
 func QueryUserInfo(userId uint) (*UserInfo, error) {
